Migrated Typesafe Config facade into separate dependency project (Resolves https://github.com/akka-js/shocon/issues/22)
