+++ facade/js/src/test/scala/AkkaConf.scala
